feat: unit tests for userRepository were rewritten, moved files to correct subdirectory
