+++ MyMusic/src/main/java/com/ciandt/summit/bootcamp2022/domains/users/User.java
@@ -62,4 +62,14 @@ public class User {
     public void setUserType(UserType userType) {
         this.userType = userType;
     }
+
+    @Override
+    public String toString() {
+        return "User{" +
+                "id='" + id + '\'' +
+                ", name='" + name + '\'' +
+                ", playlist=" + playlist +
+                ", userType=" + userType +
+                '}';
+    }
 }
