+++ MyMusic/src/main/java/com/ciandt/summit/bootcamp2022/infra/adapters/repositories/UserRepository.java
@@ -14,6 +14,7 @@ public class UserRepository implements UserRepositoryPort {
 
     @Autowired
     private SpringUserRepository userRepository;
+
     @Override
     public User findById(String id) throws UserNotFoundException {
         Optional<UserEntity> userEntity = this.userRepository.findById(id);
