+++ MyMusic/src/test/java/com/ciandt/summit/bootcamp2022/unit/repositories/UserRepositoryTest.java
@@ -1,12 +1,15 @@
-package com.ciandt.summit.bootcamp2022.unit;
+package com.ciandt.summit.bootcamp2022.unit.repositories;
 
 import com.ciandt.summit.bootcamp2022.SummitBootcampApplication;
 import com.ciandt.summit.bootcamp2022.domains.exceptions.playlists.PlaylistsNotFoundException;
 import com.ciandt.summit.bootcamp2022.domains.exceptions.users.UserNotFoundException;
+import com.ciandt.summit.bootcamp2022.domains.userType.UserType;
 import com.ciandt.summit.bootcamp2022.domains.users.User;
 import com.ciandt.summit.bootcamp2022.domains.users.dto.UserDTO;
 import com.ciandt.summit.bootcamp2022.domains.users.ports.repositories.UserRepositoryPort;
+import com.ciandt.summit.bootcamp2022.infra.adapters.entities.PlaylistEntity;
 import com.ciandt.summit.bootcamp2022.infra.adapters.entities.UserEntity;
+import com.ciandt.summit.bootcamp2022.infra.adapters.entities.UserTypeEntity;
 import com.ciandt.summit.bootcamp2022.infra.adapters.repositories.SpringUserRepository;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
@@ -16,6 +19,7 @@ import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.test.context.junit.jupiter.SpringExtension;
 
+import java.util.ArrayList;
 import java.util.Optional;
 
 import static org.junit.jupiter.api.Assertions.*;
@@ -27,7 +31,7 @@ public class UserRepositoryTest {
     @MockBean
     private SpringUserRepository springUserRepository;
 
-    @MockBean
+    @Autowired
     private UserRepositoryPort userRepositoryPort;
 
     private UserEntity userEntity;
@@ -38,17 +42,20 @@ public class UserRepositoryTest {
 
     @BeforeEach
     public void setup(){
-        userEntity = new UserEntity();
-        userEntity.setId(ID);
-        userEntity.setName("Antony");
+        PlaylistEntity playlistEntity = new PlaylistEntity("id", new ArrayList<>());
+        UserTypeEntity userTypeEntity = new UserTypeEntity("id", "free user");
+        userEntity = new UserEntity(ID, "Antony", playlistEntity, userTypeEntity);
+
+        user = new User(ID, "Antony", playlistEntity.toPlaylist(), userTypeEntity.toUserType());
     }
 
     @Test
     public void UserFound() throws UserNotFoundException {
         when(springUserRepository.findById(ID))
                 .thenReturn(Optional.of(userEntity));
+
         User result = userRepositoryPort.findById(ID);
-        assertEquals(result, user);
+        assertEquals(user.toString(), result.toString());
     }
 
     @Test
@@ -57,6 +64,11 @@ public class UserRepositoryTest {
 
         when(springUserRepository.findById(ID))
                 .thenReturn(Optional.empty());
-        assertNull(userRepositoryPort.findById(ID));
+
+        UserNotFoundException exception = assertThrows(UserNotFoundException.class, () -> {
+            userRepositoryPort.findById(ID);
+        });
+
+        assertEquals(exception.getMessage(), expectedExceptionMessage);
     }
 }
