+++ MyMusic/src/test/java/com/ciandt/summit/bootcamp2022/unit/services/UserServiceTest.java
@@ -1,4 +1,4 @@
-package com.ciandt.summit.bootcamp2022.unit;
+package com.ciandt.summit.bootcamp2022.unit.services;
 
 import com.ciandt.summit.bootcamp2022.SummitBootcampApplication;
 import com.ciandt.summit.bootcamp2022.domains.exceptions.users.UserNotFoundException;
